requirements.txt
python-dotenv
disnake
asyncpraw
